Move JdbcClientQueryEventHandler
