--- modules/client/src/main/java/org/apache/ignite/internal/client/sql/ClientSql.java
@@ -1,49 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.ignite.internal.client.sql;
-
-import org.apache.ignite.sql.IgniteSql;
-import org.apache.ignite.sql.Session;
-import org.apache.ignite.sql.Session.SessionBuilder;
-import org.apache.ignite.sql.Statement;
-import org.apache.ignite.sql.Statement.StatementBuilder;
-
-public class ClientSql implements IgniteSql {
-    @Override
-    public Session createSession() {
-        // Session and statement are just a set of defaults.
-        // All methods here are local - no network requests.
-        // Actual network calls are ony in Session.Execute()
-        return null;
-    }
-
-    @Override
-    public SessionBuilder sessionBuilder() {
-        return null;
-    }
-
-    @Override
-    public Statement createStatement(String query) {
-        return null;
-    }
-
-    @Override
-    public StatementBuilder statementBuilder() {
-        return null;
-    }
-}
